requirements.txt
# requirements.txt

Flask
gunicorn
pandas
requests
PyYAML